+++ indico/modules/events/abstracts/controllers/abstract.py
@@ -52,7 +52,7 @@ class RHEditAbstract(RHAbstractBase):
         abstract_form_class = make_abstract_form(self.event, management=self.management)
         custom_field_values = {'custom_{}'.format(x.contribution_field_id): x.data for x in self.abstract.field_values}
         defaults = FormDefaults(self.abstract, attachments=self.abstract.files, **custom_field_values)
-        form = abstract_form_class(obj=defaults, abstract=self.abstract, event=self.event)
+        form = abstract_form_class(obj=defaults, abstract=self.abstract, event=self.event, management=self.management)
         if form.validate_on_submit():
             update_abstract(self.abstract, *get_field_values(form.data))
             flash(_("Abstract modified successfully"), 'success')