+++ indico/modules/events/abstracts/controllers/display.py
@@ -69,7 +69,7 @@ class RHSubmitAbstract(RHAbstractsBase):
 
     def _process(self):
         abstract_form_class = make_abstract_form(self.event, management=self.management)
-        form = abstract_form_class(event=self.event)
+        form = abstract_form_class(event=self.event, management=self.management)
         if form.validate_on_submit():
             abstract = create_abstract(self.event, *get_field_values(form.data), send_notifications=True)
             flash(_("Your abstract '{}' has been successfully submitted. It is registered with the number "